utils, testing: Add an implementation for naf
Non-adjacent form https://en.wikipedia.org/wiki/Non-adjacent_form
of a number is a unique signed-digit representation, in which
non-zero values cannot be adjacent. For example:

(1 0 0 −1)_2 = 8 − 1 = 7

The main benefit of it is that the Hamming weight of the value will
be minimal.

The properties of NAF make it useful in various algorithms, especially
some in cryptography; e.g., for reducing the number of multiplications
needed for performing an exponentiation.

We would need the NAF represenation as part of the final exponentiation
required as part of pairing check.

Signed-off-by: Jinank Jain <jinank94@gmail.com>